requirements.txt
flask
pillow
requests
filestack-python
google-generativeai
gunicorn 
moviepy
opencv-python
pymongo
pytelegrambotapi
